Run APK tests on the main thread.
Many tests rely on being run on the main thread. In this CL we post a delayed task to run tests on the main
thread and do not block in onCreate(). This can cause ANR if the main test activity receives a touch event, but this should not happen.
BUG=125059

TEST=


Review URL: https://chromiumcodereview.appspot.com/10663024

git-svn-id: https://src.chromium.org/chrome/trunk/src/testing@144038 4ff67af0-8c30-449e-8e8b-ad334ec8d88c